fixes for dev merge
