+++ features/aave/getCommonPartsFromAppContext.ts
@@ -10,12 +10,7 @@ import { observe } from '../../blockchain/calls/observe'
 import { UserDpmAccount } from '../../blockchain/userDpmProxies'
 import { AppContext } from '../../components/AppContext'
 import { getAllowanceStateMachine } from '../stateMachines/allowance'
-import {
-  getCreateDPMAccountTransactionMachine,
-  getDPMAccountStateMachine,
-} from '../stateMachines/dpmAccount'
-import { getGasEstimation$, getOpenProxyStateMachine } from '../stateMachines/proxy/pipelines'
-import { transactionContextService } from '../stateMachines/transaction'
+import { getOpenProxyStateMachine } from '../stateMachines/proxy/pipelines'
 import { getAvailableDPMProxy$ } from './common/services/getAvailableDPMProxy'
 import { getOperationExecutorTransactionMachine } from './common/services/getTransactionMachine'
 import { getProxiesRelatedWithPosition$ } from './helpers/getProxiesRelatedWithPosition'
@@ -25,19 +20,17 @@ export function getCommonPartsFromAppContext({
   onEveryBlock$,
   connectedContext$,
   context$,
-  gasPrice$,
-  daiEthTokenPrice$,
+  gasEstimation$,
   proxyAddress$,
   userDpmProxy$,
   allowance$,
   txHelpers$,
   proxyConsumed$,
   userDpmProxies$,
+  commonTransactionServices,
+  dpmAccountStateMachine,
+  contextForAddress$,
 }: AppContext) {
-  const contextForAddress$ = connectedContext$.pipe(
-    distinctUntilKeyChanged('account'),
-    shareReplay(1),
-  )
   const disconnectedGraphQLClient$ = context$.pipe(
     distinctUntilKeyChanged('cacheApi'),
     map(({ cacheApi }) => new GraphQLClient(cacheApi)),
@@ -46,8 +39,6 @@ export function getCommonPartsFromAppContext({
   const proxyForAccount$: Observable<string | undefined> = contextForAddress$.pipe(
     switchMap(({ account }) => proxyAddress$(account)),
   )
-
-  const gasEstimation$ = curry(getGasEstimation$)(gasPrice$, daiEthTokenPrice$)
 
   const proxiesRelatedWithPosition$ = memoize(
     curry(getProxiesRelatedWithPosition$)(proxyAddress$, userDpmProxy$),
@@ -60,8 +51,6 @@ export function getCommonPartsFromAppContext({
     (token, spender) => `${token}-${spender}`,
   )
 
-  const commonTransactionServices = transactionContextService(context$)
-
   const allowanceStateMachine = getAllowanceStateMachine(
     txHelpers$,
     connectedContext$,
@@ -73,17 +62,6 @@ export function getCommonPartsFromAppContext({
     txHelpers$,
     proxyForAccount$,
     gasEstimation$,
-  )
-
-  const dpmAccountTransactionMachine = getCreateDPMAccountTransactionMachine(
-    txHelpers$,
-    connectedContext$,
-    commonTransactionServices,
-  )
-  const dpmAccountStateMachine = getDPMAccountStateMachine(
-    txHelpers$,
-    gasEstimation$,
-    dpmAccountTransactionMachine,
   )
 
   const operationExecutorTransactionMachine = curry(getOperationExecutorTransactionMachine)(
